* develop: (21 commits)
  avoid computing unused ekids
  skip if tables is empty
  fix another typo
  handle empty
  fix typo
  format
  add bulk writes to improve perf on adding org datasets
  keep check
  shuttle compile
  plus mechanic
  grant read on org metadata entity sets to org principal
  can't delete org db while connected to it
  fix org deletion logic
  make filter fields optional to support paging-only without filters
  ehh
  success!
  fix!
  cleanup
  more iteration
  iteration
  ...

# Conflicts:
#	mechanic

+++ datastore/src/main/java/com/openlattice/controllers/OrganizationsController.java
@@ -23,6 +23,8 @@ package com.openlattice.controllers;
 import com.auth0.json.mgmt.users.User;
 import com.codahale.metrics.annotation.Timed;
 import com.google.common.base.Preconditions;
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
 import com.openlattice.apps.services.AppService;
@@ -168,15 +170,12 @@ public class OrganizationsController implements AuthorizingComponent, Organizati
     @DeleteMapping( value = ID_PATH, produces = MediaType.APPLICATION_JSON_VALUE )
     @ResponseStatus( HttpStatus.OK )
     public Void destroyOrganization( @PathVariable( ID ) UUID organizationId ) {
-        AclKey aclKey = ensureOwner( organizationId );
-
+        ensureOwner( organizationId );
         ensureObjectCanBeDeleted( organizationId );
         organizations.ensureOrganizationExists( organizationId );
 
         organizations.destroyOrganization( organizationId );
         edms.deleteOrganizationExternalDatabase( organizationId );
-        authorizations.deletePermissions( aclKey );
-        securableObjectTypes.deleteSecurableObjectType( new AclKey( organizationId ) );
         return null;
     }
 
@@ -474,7 +473,7 @@ public class OrganizationsController implements AuthorizingComponent, Organizati
             consumes = MediaType.APPLICATION_JSON_VALUE,
             produces = MediaType.APPLICATION_JSON_VALUE
     )
-        public Map<UUID, Integer> getMemberCountForOrganizations( Set<UUID> organizationIds ) {
+    public Map<UUID, Integer> getMemberCountForOrganizations( Set<UUID> organizationIds ) {
         EnumSet<Permission> readPermissions = EnumSet.of( Permission.READ );
         accessCheck( organizationIds.stream().collect( Collectors.toMap( AclKey::new, id -> readPermissions ) ) );
 
@@ -725,12 +724,13 @@ public class OrganizationsController implements AuthorizingComponent, Organizati
                 .getExternalDatabaseTablesWithColumns( organizationId );
         tables
                 .forEach( ( t, c ) -> {
-                    organizationMetadataEntitySetsService.addDataset( organizationId, t.getSecond() );
                     organizationMetadataEntitySetsService
-                            .addDatasetColumns(
+                            .addDatasetsAndColumns(
                                     organizationId,
-                                    t.getSecond(),
-                                    c.stream().map( Entry::getValue ).collect( Collectors.toList() ) );
+                                    ImmutableList.of( t.getSecond() ),
+                                    ImmutableMap.of( t.getSecond().getId(),
+                                            c.stream().map( Entry::getValue ).collect( Collectors.toList() ) ) );
+
                 } );
 
         entitySetManager
@@ -739,8 +739,9 @@ public class OrganizationsController implements AuthorizingComponent, Organizati
                     var entitySet = Objects.requireNonNull( entitySetManager.getEntitySet( e ) );
                     var propertyTypes = edmService.getPropertyTypesOfEntityType( entitySet.getEntityTypeId() );
 
-                    organizationMetadataEntitySetsService.addDataset( entitySet );
-                    organizationMetadataEntitySetsService.addDatasetColumns( entitySet, propertyTypes.values() );
+                    organizationMetadataEntitySetsService.addDatasetsAndColumns(
+                            ImmutableList.of( entitySet ),
+                            ImmutableMap.of( entitySet.getId(), propertyTypes.values() ) );
                 } );
 
         return null;
